re-shuffle fns and add builder selection validation
